requirements.txt
ultralytics  # Critical for YOLOv8 functionality
streamlit
numpy
opencv-python
Pillow